requirements.txt
aiohappyeyeballs==2.6.1 ; python_full_version == "3.11.6"
aiohttp-cors==0.8.1 ; python_full_version == "3.11.6"
aiohttp==3.13.2 ; python_full_version == "3.11.6"
aiosignal==1.4.0 ; python_full_version == "3.11.6"
annotated-doc==0.0.3 ; python_full_version == "3.11.6"
annotated-types==0.7.0 ; python_full_version == "3.11.6"
anyio==4.11.0 ; python_full_version == "3.11.6"
attrs==25.4.0 ; python_full_version == "3.11.6"
bcrypt==5.0.0 ; python_full_version == "3.11.6"
cachetools==6.2.1 ; python_full_version == "3.11.6"
certifi==2025.10.5 ; python_full_version == "3.11.6"
cffi==2.0.0 ; python_full_version == "3.11.6" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4 ; python_full_version == "3.11.6"
click==8.2.1 ; python_full_version == "3.11.6"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_full_version == "3.11.6"
colorful==0.5.8 ; python_full_version == "3.11.6"
cryptography==46.0.3 ; python_full_version == "3.11.6"
dill==0.3.9 ; python_full_version == "3.11.6"
distlib==0.4.0 ; python_full_version == "3.11.6"
fastapi==0.121.0 ; python_full_version == "3.11.6"
filelock==3.20.0 ; python_full_version == "3.11.6"
frozenlist==1.8.0 ; python_full_version == "3.11.6"
fsspec==2025.10.0 ; python_full_version == "3.11.6"
google-api-core==2.28.1 ; python_full_version == "3.11.6"
google-auth==2.43.0 ; python_full_version == "3.11.6"
googleapis-common-protos==1.71.0 ; python_full_version == "3.11.6"
grpcio==1.76.0 ; python_full_version == "3.11.6"
h11==0.16.0 ; python_full_version == "3.11.6"
httptools==0.7.1 ; python_full_version == "3.11.6"
idna==3.11 ; python_full_version == "3.11.6"
importlib-metadata==8.7.0 ; python_full_version == "3.11.6"
jsonschema-specifications==2025.9.1 ; python_full_version == "3.11.6"
jsonschema==4.25.1 ; python_full_version == "3.11.6"
mako==1.3.10 ; python_full_version == "3.11.6"
markdown==3.10 ; python_full_version == "3.11.6"
markupsafe==3.0.3 ; python_full_version == "3.11.6"
msgpack==1.1.2 ; python_full_version == "3.11.6"
multidict==6.7.0 ; python_full_version == "3.11.6"
numpy==2.3.4 ; python_full_version == "3.11.6"
opencensus-context==0.1.3 ; python_full_version == "3.11.6"
opencensus==0.11.4 ; python_full_version == "3.11.6"
opentelemetry-api==1.38.0 ; python_full_version == "3.11.6"
opentelemetry-exporter-prometheus==0.59b0 ; python_full_version == "3.11.6"
opentelemetry-proto==1.38.0 ; python_full_version == "3.11.6"
opentelemetry-sdk==1.38.0 ; python_full_version == "3.11.6"
opentelemetry-semantic-conventions==0.59b0 ; python_full_version == "3.11.6"
packaging==25.0 ; python_full_version == "3.11.6"
pandas==2.3.3 ; python_full_version == "3.11.6"
paramiko==3.5.1 ; python_full_version == "3.11.6"
pdoc3==0.10.0 ; python_full_version == "3.11.6"
pip-chill==1.0.3 ; python_full_version == "3.11.6"
platformdirs==4.5.0 ; python_full_version == "3.11.6"
prometheus-client==0.23.1 ; python_full_version == "3.11.6"
propcache==0.4.1 ; python_full_version == "3.11.6"
proto-plus==1.26.1 ; python_full_version == "3.11.6"
protobuf==6.33.0 ; python_full_version == "3.11.6"
py-spy==0.4.1 ; python_full_version == "3.11.6"
pyarrow==22.0.0 ; python_full_version == "3.11.6"
pyasn1-modules==0.4.2 ; python_full_version == "3.11.6"
pyasn1==0.6.1 ; python_full_version == "3.11.6"
pycparser==2.23 ; platform_python_implementation != "PyPy" and implementation_name != "PyPy" and python_full_version == "3.11.6"
pydantic-core==2.41.5 ; python_full_version == "3.11.6"
pydantic==2.12.4 ; python_full_version == "3.11.6"
pynacl==1.6.0 ; python_full_version == "3.11.6"
python-dateutil==2.9.0.post0 ; python_full_version == "3.11.6"
python-dotenv==1.2.1 ; python_full_version == "3.11.6"
pytz==2025.2 ; python_full_version == "3.11.6"
pyyaml==6.0.3 ; python_full_version == "3.11.6"
ray==2.51.1 ; python_full_version == "3.11.6"
referencing==0.37.0 ; python_full_version == "3.11.6"
requests==2.32.5 ; python_full_version == "3.11.6"
rpds-py==0.28.0 ; python_full_version == "3.11.6"
rsa==4.9.1 ; python_full_version == "3.11.6"
six==1.17.0 ; python_full_version == "3.11.6"
smart-open==7.4.4 ; python_full_version == "3.11.6"
sniffio==1.3.1 ; python_full_version == "3.11.6"
starlette==0.49.3 ; python_full_version == "3.11.6"
tensorboardx==2.6.4 ; python_full_version == "3.11.6"
typing-extensions==4.15.0 ; python_full_version == "3.11.6"
typing-inspection==0.4.2 ; python_full_version == "3.11.6"
tzdata==2025.2 ; python_full_version == "3.11.6"
urllib3==2.5.0 ; python_full_version == "3.11.6"
uvicorn==0.38.0 ; python_full_version == "3.11.6"
uvloop==0.22.1 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_full_version == "3.11.6"
virtualenv==20.35.4 ; python_full_version == "3.11.6"
watchfiles==1.1.1 ; python_full_version == "3.11.6"
websockets==15.0.1 ; python_full_version == "3.11.6"
wrapt==2.0.0 ; python_full_version == "3.11.6"
yarl==1.22.0 ; python_full_version == "3.11.6"
zipp==3.23.0 ; python_full_version == "3.11.6"
